+++ Insights.Service/Interface/IAdInsightsService.cs
@@ -4,7 +4,7 @@ namespace Insights.Gateway.Services;
 
 public interface IAdInsightsService
 {
-    Task<int> GetClicksAsync(string campaignID);
-    Task<int> GetImpressionsAsync(string campaignID);
+    Task<long> GetClicksAsync(string campaignID);
+    Task<long> GetImpressionsAsync(string campaignID);
     Task<int> GetClickToBasketAsync(string campaignID);
 }
